Improve docstring
